project/mlops-project/requirements.txt
scikit-learn
pandas
mlflow
matplotlib
seaborn
pyyaml
pytest
ruff
tqdm
